Revert "Revert "add new main window and windows for adding cars/customers""
This reverts commit d4ce92f597804113be6d7177ee65aef384990bd2.

+++ CarRentalDatabase/dbo/Tables/Customer.sql
@@ -4,6 +4,6 @@
     [firstname] VARCHAR(50) NOT NULL, 
     [lastname] VARCHAR(50) NOT NULL, 
     [title] VARCHAR(50) NULL DEFAULT 'Herr/Frau', 
-    [age] SMALLINT NOT NULL, 
+    [age] INT NOT NULL, 
     [notes] VARCHAR(50) NULL DEFAULT 'N/A'
 )
